build.sh: check for existence of qmake, don't call it with wrong opts
This fixes #387.

Formerly, `qmake --help` was called, and the return value was checked
for being 0, but `--help` is not a valid option in any of my qmakes.

`-help` would be correct, but even with that, qmake returns a non-zero
value.

So, simply replaced that with a check for the existence of the qmake
executable.