Generate and deploy source tarball (Fix #2878) (#3036)
This builds a real tarball, as opposed to the flawed GitHub-generated
one, and also includes the following changes:

Individual builds now have dedicated names like "linux-x86_64". The
structure in the yml is now very similar to how it is in rizin.

Since that means builds are renamed, the filename has also been changed
from the meaningless "ccpp.yml" to "ci.yml", as that would have happened
sooner or later anyway and now will not produce additional intermediate
rename states.

The workflow name inside that file is now also just "CI" since adding
"Cutter" there is redundant.